Feature/#6

+++ src/main/java/com/com2here/com2hereback/service/UserServiceImpl.java
@@ -1,5 +1,7 @@
 package com.com2here.com2hereback.service;
 
+import com.com2here.com2hereback.common.BaseResponseStatus;
+import com.com2here.com2hereback.common.CMResponse;
 import com.com2here.com2hereback.domain.User;
 import com.com2here.com2hereback.dto.ShowUserResponseDto;
 import com.com2here.com2hereback.dto.UserRequestDto;
@@ -9,6 +11,8 @@ import com.com2here.com2hereback.security.TokenProvider;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,45 +29,93 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public int RegisterUser(UserRequestDto userRequestDto) {
+    public CMResponse RegisterUser(UserRequestDto userRequestDto) {
+        BaseResponseStatus status;
+
+        // 400 : 데이터 누락
+        if (userRequestDto == null || userRequestDto.getPassword() == null || userRequestDto.getEmail() == null) {
+            status = BaseResponseStatus.WRONG_PARAM;
+            return CMResponse.fail(status.getCode(),status,null);
+        }
+
+        // 400 : 비밀번호 형식 불일치
+        String regex = "^(?=.*[a-zA-Z])(?=.*\\d)(?=.*[!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>\\/?]).{8,20}$";
+        if (!(userRequestDto.getPassword()).matches(regex)) {
+            status = BaseResponseStatus.PASSWORD_FORMAT_INVALID;
+            return CMResponse.fail(status.getCode(),status,null);
+        }
+
+        // 중복된 이메일
+        if (userRepository.existsByEmail(userRequestDto.getEmail())) {
+            status = BaseResponseStatus.DUPLICATE_EMAIL;
+            return CMResponse.fail(status.getCode(),status,null);
+        }
 
         User user = User.builder()
             .username(userRequestDto.getUsername())
             .password(bCryptPasswordEncoder.encode(userRequestDto.getPassword()))
             .email(userRequestDto.getEmail()).build();
 
-        System.out.println(userRequestDto.getPassword());
         userRepository.save(user);
 
-        return 201;
+        status = BaseResponseStatus.REGISTRATION_SUCCESS;
+
+        return CMResponse.success(status.getCode(),status,null);
     }
 
     @Override
-    public UserResponseDto LoginUser(UserRequestDto userRequestDto) {
+    public CMResponse LoginUser(UserRequestDto userRequestDto) {
+        BaseResponseStatus status;
 
-        User user = userRepository.findByEmail(userRequestDto.getEmail());
-        if (user != null && bCryptPasswordEncoder.matches(userRequestDto.getPassword(), user.getPassword())) {
-            String token = tokenProvider.create(user);
-            return UserResponseDto.builder()
-                .message("로그인 성공")
-                .token(token)
-                .build();
+        // 400 : 데이터 누락
+        if(userRequestDto.getEmail()==null || userRequestDto.getPassword()==null){
+            status = BaseResponseStatus.WRONG_PARAM;
+            return CMResponse.fail(status.getCode(),status);
         }
-        return null;
+
+        // 회원 존재 X 2106
+        User user = userRepository.findByEmail(userRequestDto.getEmail());
+        if(user==null || !bCryptPasswordEncoder.matches(userRequestDto.getPassword(), user.getPassword())){
+            status = BaseResponseStatus.NO_EXIST_MEMBERS;
+            return CMResponse.fail(status.getCode(),status);
+        }
+        String token = tokenProvider.create(user);
+
+        status = BaseResponseStatus.LOGIN_SUCCESS;
+
+        return CMResponse.success(status.getCode(),status,token);
     }
 
     @Override
-    public ShowUserResponseDto ShowUser(HttpServletRequest request) {
-        User user = userRepository.findById(tokenProvider.getSubject(request.getHeader("Authorization")));
+    public CMResponse ShowUser(HttpServletRequest request) {
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        int userId = (int) authentication.getPrincipal(); // userId를 가져옵니다.
+
+        User user = userRepository.findById(userId);
+        if (user == null) {
+            // return BaseResponseStatus.NO_EXIST_MEMBERS; // 존재하지 않는 사용자
+            return CMResponse.fail(2106,BaseResponseStatus.NO_EXIST_MEMBERS);
+        }
         ShowUserResponseDto showUserResponseDto = ShowUserResponseDto.entityToDto(user);
-        return showUserResponseDto;
+        return CMResponse.success(200,BaseResponseStatus.USER_INFO_RETRIEVED,showUserResponseDto);
     }
 
     @Override
     @Transactional
-    public void updateUser(UserRequestDto userRequestDto, HttpServletRequest request) {
-        User user = userRepository.findById(tokenProvider.getSubject(request.getHeader("Authorization")));
+    public CMResponse updateUser(UserRequestDto userRequestDto, HttpServletRequest request) {
+        BaseResponseStatus status;
 
+        // 400 : 데이터 누락
+        if(userRequestDto==null){
+            status = BaseResponseStatus.WRONG_PARAM;
+            return CMResponse.fail(status.getCode(),status);
+        }
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        int userId = (int) authentication.getPrincipal(); // userId를 가져옵니다.
+
+        User user = userRepository.findById(userId);
         User updatedUser = User.builder()
             .user_id(user.getUser_id())
             .username(userRequestDto.getUsername() != null ? userRequestDto.getUsername() : user.getUsername())
@@ -72,22 +124,42 @@ public class UserServiceImpl implements UserService {
             .build();
 
         userRepository.save(updatedUser);
+
+        status = BaseResponseStatus.MEMBER_INFO_UPDATED;
+        return CMResponse.success(status.getCode(),status,null);
     }
 
     @Override
     @Transactional
-    public int deleteUser(UserRequestDto userRequestDto, HttpServletRequest request) {
+    public CMResponse deleteUser(UserRequestDto userRequestDto, HttpServletRequest request) {
+        BaseResponseStatus status;
 
-        User user = userRepository.findById(tokenProvider.getSubject(request.getHeader("Authorization")));
-
-        // 입력된 비밀번호와 저장된 비밀번호 비교
-        if (!bCryptPasswordEncoder.matches(userRequestDto.getPassword(), user.getPassword())) {
-            return 403;
+        // 400 : 데이터 누락
+        if(userRequestDto==null){
+            status = BaseResponseStatus.WRONG_PARAM;
+            return CMResponse.fail(status.getCode(),status);
         }
 
-        // 비밀번호가 일치하면 사용자 삭제
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        int userId = (int) authentication.getPrincipal(); // userId를 가져옵니다.
+
+        // 2106
+        User user = userRepository.findById(userId);
+        if (user == null) {
+            status =  BaseResponseStatus.NO_EXIST_MEMBERS; // 존재하지 않는 사용자
+            return CMResponse.fail(status.getCode(),status);
+        }
+        // 2202
+        if (!bCryptPasswordEncoder.matches(userRequestDto.getPassword(), user.getPassword())) {
+            status = BaseResponseStatus.PASSWORD_MISMATCH; // 비밀번호 불일치
+            return CMResponse.fail(status.getCode(),status);
+        }
+
         userRepository.delete(user);
-        return 200;
+        status = BaseResponseStatus.MEMBER_DELETE_SUCCESS;
+
+        // 200
+        return CMResponse.success(status.getCode(),status,null);
     }
